trying to add json

+++ service/src/main/java/com/iskhak/padie/dao/UserDAO.java
@@ -2,7 +2,7 @@ package com.iskhak.padie.dao;
 
 import java.util.List;
 
-import com.iskhak.padie.model.User;
+import com.iskhak.padie.model.security.User;
 
 public interface UserDAO {
 	public List<User> list();
@@ -11,7 +11,7 @@ public interface UserDAO {
 	
 	public void saveOrUpdate(User user);
 	
-	public void delete(int id);
+	public void delete(long id);
 	
 	public boolean login(User user);
 }
