+++ service/src/main/java/com/iskhak/padie/dao/UserDAOImpl.java
@@ -10,7 +10,7 @@ import org.hibernate.SessionFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
-import com.iskhak.padie.model.User;
+import com.iskhak.padie.model.security.User;
 
 @Repository
 public class UserDAOImpl implements UserDAO {
@@ -44,7 +44,7 @@ public class UserDAOImpl implements UserDAO {
  
     @Override
     @Transactional
-    public void delete(int id) {
+    public void delete(long id) {
         User userToDelete = new User();
         userToDelete.setId(id);
         sessionFactory.getCurrentSession().delete(userToDelete);
